% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HydroGOF.R
\name{GOFmeas}
\alias{GOFmeas}
\alias{print.GOFmeas}
\alias{summary.GOFmeas}
\title{The GOFmeas class}
\usage{
\method{print}{GOFmeas}(x, ...)

\method{summary}{GOFmeas}(object, ...)
}
\arguments{
\item{x}{an object of the GOF class, typically the output of a GOFmeasures call}

\item{...}{not used}

\item{object}{an object of the GOF class, typically the output of a GOFmeasures call}
}
\description{
Measures of goodness of fit

Measures of goodness of fit
}
\seealso{
\code{\link{GOFmeasures}}
}
